renamer filer og bytter ut gammel tabell med ny
