+++ src/frontend/komponenter/Oppgavebenk/OppgaveTabellUtils.tsx
